+++ server/app/views/admin/programs/ProgramBlocksView.java
@@ -50,7 +50,7 @@ import views.components.ButtonStyles;
 import views.components.FieldWithLabel;
 import views.components.Icons;
 import views.components.Modal;
-import views.components.QuestionBank;
+import views.components.ProgramQuestionBank;
 import views.components.SvgTag;
 import views.components.ToastMessage;
 import views.style.AdminStyles;
@@ -187,7 +187,7 @@ public final class ProgramBlocksView extends ProgramBaseView {
                   programDefinition,
                   blockDefinition,
                   csrfTag,
-                  QuestionBank.shouldShowQuestionBank(request),
+                  ProgramQuestionBank.shouldShowQuestionBank(request),
                   settingsManifest.getPhoneQuestionTypeEnabled(request)))
           .addMainContent(addFormEndpoints(csrfTag, programDefinition.id(), blockId))
           .addModals(blockDescriptionEditModal, blockDeleteScreenModal);
@@ -992,7 +992,7 @@ public final class ProgramBlocksView extends ProgramBaseView {
       ProgramDefinition program,
       BlockDefinition blockDefinition,
       InputTag csrfTag,
-      QuestionBank.Visibility questionBankVisibility,
+      ProgramQuestionBank.Visibility questionBankVisibility,
       boolean phoneQuestionTypeEnabled) {
     String addQuestionAction =
         controllers.admin.routes.AdminProgramBlockQuestionsController.create(
@@ -1000,13 +1000,13 @@ public final class ProgramBlocksView extends ProgramBaseView {
             .url();
 
     String redirectUrl =
-        QuestionBank.addShowQuestionBankParam(
+        ProgramQuestionBank.addShowQuestionBankParam(
             controllers.admin.routes.AdminProgramBlocksController.edit(
                     program.id(), blockDefinition.id())
                 .url());
-    QuestionBank qb =
-        new QuestionBank(
-            QuestionBank.QuestionBankParams.builder()
+    ProgramQuestionBank qb =
+        new ProgramQuestionBank(
+            ProgramQuestionBank.ProgramQuestionBankParams.builder()
                 .setQuestionAction(addQuestionAction)
                 .setCsrfTag(csrfTag)
                 .setQuestions(questionDefinitions)
